+++ xrpl/time/conversion_test.go
@@ -1,4 +1,4 @@
-package xrpl
+package time
 
 import (
 	"testing"
